formatted output and bug fixes

+++ DHT.h
@@ -3,6 +3,9 @@
 #ifndef DHT_H
 #define DHT_H
 
+#include <ctime>
+#include <chrono>
+#include <thread>
 #include "CIrcularLL.h"
 #include "Machine.h"
 #include "Utils.h"
@@ -36,6 +39,9 @@ public:
         else return false;
     }
 	void makeRoutingTables() {
+        if (ring.isEmpty()) return;
+        cout << "Updating Routing Tables"; this_thread::sleep_for(chrono::seconds(1)); cout << "."; this_thread::sleep_for(chrono::seconds(1));
+        cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << "." << endl;
 		CircleListNode<Machine>* start = ring.head;
         if (start->next == ring.head) {
             for (int i = 0; i < identifierSpace; i++) {
@@ -86,9 +92,13 @@ public:
 			}
 			start->data.addRoutingTableEntry(&(probe->data));
 		}
-
+        cout << "Routing Tables Updated Successfully" << endl;
 	}
     void showMachines() {
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl;
+            return;
+        }
         std::string dotCode = "graph CircularLinkedList {\n";
         CircleListNode<Machine>* current = ring.head;
 
@@ -162,7 +172,7 @@ public:
         // we create the directory for this machine and then shift the files
         handle.createFolder("D:\\storage\\DHT\\" + machineName);
         current->data.shiftFiles('i', current->next->data);
-        
+        cout << "Machine Inserted With Name: " << machineName << endl; cout << "Id Assigned To This Machine: " << machineId.getBIG_INT() << endl;
     }
     void removeMachine(BIG_INT id) { // implementation to be done.
         if (ring.isEmpty()) {
@@ -200,6 +210,10 @@ public:
     }
 
     Machine* routerSearch(BIG_INT& e, BIG_INT& p) {
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl;
+            return nullptr;
+        }
         if (!e.validate() || e.getBIG_INT()[0] == '-' || e >= power(BIG_INT("2"), identifierSpace)) return nullptr;
         else if (!p.validate() || p.getBIG_INT()[0] == '-' || p >= power(BIG_INT("2"), identifierSpace)) return nullptr;
         if (ring.head == ring.head->next) {
@@ -207,44 +221,65 @@ public:
                 return &ring.head->data;
             }
             else {
-                cout << "machine does not exist" << endl;
+                cout << "Machine Does Not Exist" << endl;
                 return nullptr;
             }
         }
         // first step is to find the machine from which the request has to be generated.
+        cout << "Finding The Origin Machine"; this_thread::sleep_for(chrono::seconds(1)); cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << ".";
+        this_thread::sleep_for(chrono::seconds(1)); cout << "." << endl;
         Machine* currentMachinePtr = &ring.head->data;
+        cout << "Control Passed To The Head Machine: " << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << endl;
         while (!(p == currentMachinePtr->getId())) {
+            this_thread::sleep_for(chrono::seconds(1));
             if (p < currentMachinePtr->getId() && currentMachinePtr == &ring.head->data) {
                 break;
             }
             else if (p > currentMachinePtr->getId() && p <= currentMachinePtr->getRoutingTable().front()->getId()) {
+                cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                 currentMachinePtr = currentMachinePtr->getRoutingTable().front();
+                cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 break;
             }
             else if (p > currentMachinePtr->getId() && currentMachinePtr->getRoutingTable().front()->getId() < currentMachinePtr->getId()) { // case where next is head
+                cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                 currentMachinePtr = currentMachinePtr->getRoutingTable().front();
+                cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 break;
             }
             else if (p < currentMachinePtr->getId() && currentMachinePtr->getRoutingTable().front()->getId() < currentMachinePtr->getId()) { // case where next is head
+                cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                 currentMachinePtr = currentMachinePtr->getRoutingTable().front();
+                cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 break;
             }
             else {
+                cout << "Traversing The Routing Table Of Machine: [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 for (DoublyLinkedList<Machine*>::Iterator it = currentMachinePtr->getRoutingTable().begin(); it != currentMachinePtr->getRoutingTable().end(); ++it) {
                     if (it + 1 == nullptr) {
-                        if (currentMachinePtr->getId() == (*it)->getId())
+                        if (currentMachinePtr->getId() == (*it)->getId()) {
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = currentMachinePtr->getRoutingTable().front();
-                        else
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
+                        }
+                        else {
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = *it;
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
+                        }
                         break;
                     }
                     else {
                         if (p > (*it)->getId() && (*it)->getId() > (it + 1)->data->getId()) { // the case the where the next machine is the head
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = *it;
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                             break;
                         }
                         if (p > (*it)->getId() && p <= (it + 1)->data->getId()) {
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = *it;
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                             break;
                         }
                     }
@@ -252,11 +287,15 @@ public:
             }
         }
         if (!(currentMachinePtr->getId() == p)) {
-            cout << "Machine does not exist" << endl; return nullptr;
+            cout << "Machine Does Not Exist" << endl; return nullptr;
         }
+        cout << "Origin Machine Found: [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
+        cout << "Starting Search For The File With Id: " << e << " ."; this_thread::sleep_for(chrono::seconds(1));
+        cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << "." << endl;
         // now we have the pointer to the machine from which the request has to be generated.
         p = currentMachinePtr->getId();
         while (true) { // search the ring dht using the machine's routing tables
+            this_thread::sleep_for(chrono::seconds(1));
             //if (!(e == p) && (e < p) && (e < currentMachinePtr->getRoutingTable().front()->getId())) { // this means we have found the machine
             //    break;
             //}
@@ -267,27 +306,38 @@ public:
                 break;
             }
             else if (e > p && e <= currentMachinePtr->getRoutingTable().front()->getId()) { // the next machine contains the file
+                cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                 currentMachinePtr = currentMachinePtr->getRoutingTable().front();
+                cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 break;
             }
             else if ((p > currentMachinePtr->getRoutingTable().front()->getId()) && (e > p || e <= currentMachinePtr->getRoutingTable().front()->getId())) { // the case where the next machine is the head
                 // here we consider two cases. one in which e is greater than current. and another where e is less than or equal to the head
+                cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                 currentMachinePtr = currentMachinePtr->getRoutingTable().front();
+                cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 break;
             }
             else {
+                cout << "Traversing The Routing Table Of Machine: [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                 for (DoublyLinkedList<Machine*>::Iterator it = currentMachinePtr->getRoutingTable().begin(); it != currentMachinePtr->getRoutingTable().end(); ++it) {
                     if (it + 1 == nullptr) { // not found in the routing table. we point to the last entry in the routing table
+                        cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                         currentMachinePtr = *it;
+                        cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                         break;
                     }
                     else {
                         if (e > (*it)->getId() && (*it)->getId() > (it + 1)->data->getId()) { // the case the where the next machine is the head
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = *it;
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                             break;
                         }
                         if (e > (*it)->getId() && e <= (it + 1)->data->getId()) { // file found in this range
+                            cout << "Control Passed From [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "] To [";
                             currentMachinePtr = *it;
+                            cout << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
                             break;
                         }
                     }
@@ -295,6 +345,7 @@ public:
             }
             p = currentMachinePtr->getId();
         }
+        cout << "Machine Found Which Has This File: [" << currentMachinePtr->getName() << "_" << currentMachinePtr->getId() << "]" << endl;
         // till here we have reached the machine where the actual file is stored
         return currentMachinePtr;
     }
@@ -302,47 +353,63 @@ public:
     void searchFile(BIG_INT& e, BIG_INT& p) { // here we implement the search algorithm using the routing tables.
         Machine* machine = routerSearch(e, p);
         if (machine == nullptr) {
-            cout << "no machine with id: " << p.getBIG_INT() << endl;
+            cout << "No Machine With Id: " << p.getBIG_INT() << endl;
             return;
         }
+        this_thread::sleep_for(chrono::seconds(1));
+        cout << "Going In The BTree Of Machine [" << machine->getName() << "_" << machine->getId() << "] .";
+        this_thread::sleep_for(chrono::seconds(1)); cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << "." << endl;
         Pair<BTreeNode<KeyValuePair<BIG_INT, LinkedList<string>>>*, int, int> searchResult = machine->searchFile(e);
         if (searchResult.first == nullptr) {
-            cout << "No such file in the system" << endl;
+            cout << "No Such File In The System" << endl;
         }
         else {
             int count = 0;
-            cout << "The files with id " << e.getBIG_INT() << " are: " << endl;
+            cout << "The Files With Id " << e.getBIG_INT() << " are: " << endl;
             for (LinkedList<string>::Iterator it = searchResult.first->keys[searchResult.second].value.begin(); it != searchResult.first->keys[searchResult.second].value.end(); ++it) {
                 cout << ++count << ". " << *it << endl;
             }
         }
     }
     void addFile(const string& path) {
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl; return;
+        }
+        cout << "Starting To Add File [" << path << "] To The System.";
+        this_thread::sleep_for(chrono::seconds(1)); cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << '.' << endl;
         ifstream file(path);
         if (file.is_open()) {
             BIG_INT p = ring.head->data.getId();
             string hash = SHA1::from_file(path);
             BIG_INT e = hashMod(hash, identifierSpace);
-            cout << "File assigned the id of: " << e.getBIG_INT() << endl;
+            cout << "File Assigned The Id Of: " << e.getBIG_INT() << endl;
             Machine* machine = routerSearch(e, p);
-            machine->addFile(e, ext(path), path);
-            cout << "file inserted in system with id: " << e.getBIG_INT() << endl;
+            if (machine) {
+                machine->addFile(e, ext(path), path);
+                cout << "File Inserted In the System With Id: " << e.getBIG_INT() << endl;
+            }
+            else {
+                cout << "File Not Inserted" << endl;
+            }
             file.close();
         }
         else {
-            cout << "no file exists with this path" << endl;
+            cout << "No File Exists With This Path" << endl;
         }
         
     }
     void printRoutingTable() {
-        string id; cout << "Enter the id of the machine you want to print the routing table of: "; getline(cin, id);
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl; return;
+        }
+        string id; cout << "Enter The Id Of The Machine You Want To Print The Routing Table Of: "; getline(cin, id);
         BIG_INT machineId(id);
         while (validateId(machineId)) {
             if (ring.exists(Machine(id))) {
                 break;
             }
             else {
-                cout << "Invalid Id, enter again: ";
+                cout << "Invalid Id, Enter Again: ";
             }
             getline(cin, id);
             machineId = BIG_INT(id);
@@ -354,14 +421,17 @@ public:
         }
     }
     void printBtree() {
-        string id; cout << "Enter the id of the machine you want to print the B-Tree of: "; getline(cin, id);
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl; return;
+        }
+        string id; cout << "Enter The Id Of The Machine You Want To Print The B-Tree Of: "; getline(cin, id);
         BIG_INT machineId(id);
         while (validateId(machineId)) {
             if (ring.exists(Machine(id))) {
                 break;
             }
             else {
-                cout << "Invalid Id, enter again: ";
+                cout << "Invalid Id, Enter Again: ";
             }
             getline(cin, id);
             machineId = BIG_INT(id);
@@ -373,10 +443,14 @@ public:
         }
     }
     void removeFile(BIG_INT e) {
+        if (ring.isEmpty()) {
+            cout << "DHT Is Empty" << endl; return;
+        }
+        cout << "Starting To Remove The File With Id: " << e; this_thread::sleep_for(chrono::seconds(1)); cout << ".";
+        this_thread::sleep_for(chrono::seconds(1)); cout << "."; this_thread::sleep_for(chrono::seconds(1)); cout << "." << endl;
         BIG_INT p = ring.head->data.getId();
         Machine* machine = routerSearch(e, p);
         machine->removeFile(e);
-        cout << "file removed" << endl;
     }
     ~DHT() {
         this->ring.makenull();
